+++ clease/datastructures/__init__.py
@@ -2,7 +2,14 @@
 
 from .system_changes import *
 from .four_vector import *
+from .trans_matrix import *
 from .figure import *
 
 ADDITIONAL = ("system_changes",)
-__all__ = system_changes.__all__ + four_vector.__all__ + figure.__all__ + ADDITIONAL
+__all__ = (
+    system_changes.__all__
+    + four_vector.__all__
+    + figure.__all__
+    + trans_matrix.__all__
+    + ADDITIONAL
+)